+++ packages/e2e/cypress/e2e/dates.cy.ts
@@ -89,12 +89,14 @@ describe('Date tests', () => {
             'order date month{enter}',
         );
         cy.contains('button', 'Select a date').click();
-        cy.findByRole('dialog').within(() => {
-            cy.contains('button', new Date().getFullYear()).click();
-            cy.get('button').find('[data-previous="true"]').click();
-            cy.contains('button', 2018).click();
-            cy.contains('button', 'Feb').click();
-        });
+        cy.findAllByRole('dialog')
+            .eq(1)
+            .within(() => {
+                cy.contains('button', new Date().getFullYear()).click();
+                cy.get('button').find('[data-previous="true"]').click();
+                cy.contains('button', 2018).click();
+                cy.contains('button', 'Feb').click();
+            });
 
         cy.contains('Apply').click();